+++ platform/app/src/routes/index.tsx
@@ -38,9 +38,14 @@ const NotFoundStudy = () => {
   return (
     <div className="text-foreground absolute flex h-full w-full items-center justify-center">
       <div>
-        <h4>One or more of the requested studies are not available at this time.</h4>
+        <h4 data-cy="study-not-found-message">
+          One or more of the requested studies are not available at this time.
+        </h4>
         {showStudyList && (
-          <p className="mt-2">
+          <p
+            className="mt-2"
+            data-cy="return-to-study-list-message"
+          >
             Return to the{' '}
             <Link
               className="text-highlight"